+++ DigitaleBriefwahl/ExceptionHandling/ExceptionLoggingUI.cs
@@ -1,11 +1,11 @@
-// Copyright (c) 2017 Eberhard Beilharz
+// Copyright (c) 2017-2018 Eberhard Beilharz
 // This software is licensed under the GNU General Public License version 3
 // (https://opensource.org/licenses/GPL-3.0)
 
-using System;
 using System.Runtime.CompilerServices;
-using Bugsnag;
+using Bugsnag.Payload;
 using Eto.Forms;
+using Exception = System.Exception;
 
 namespace DigitaleBriefwahl.ExceptionHandling
 {
@@ -13,7 +13,7 @@ namespace DigitaleBriefwahl.ExceptionHandling
 	{
 		private Control Parent { get; }
 
-		internal ExceptionLoggingUI(string apiKey, string appName, string[] args,
+		private ExceptionLoggingUI(string apiKey, string appName, string[] args,
 			string launcherVersion, object parent, string callerFilePath)
 			: base(apiKey, appName, args, launcherVersion, callerFilePath)
 		{
@@ -24,25 +24,30 @@ namespace DigitaleBriefwahl.ExceptionHandling
 		private void OnUnhandledInstanceException(object sender, Eto.UnhandledExceptionEventArgs e)
 		{
 			var ex = e.ExceptionObject as Exception;
-			var errorReport = new ErrorReport(ex);
-			if (errorReport.ShowModal(Parent))
-			{
-				if (ex != null)
-				{
-					Config.Metadata.AddToTab("Exception Details",
-						"stacktrace", ex.StackTrace);
-					Notify(ex, Severity.Error);
-				}
-			}
-			Application.Instance.Quit();
+			Notify(ex, HandledState.ForUnhandledException());
+			if (e.IsTerminating)
+				Application.Instance.Quit();
 		}
 
-		public static ExceptionLogging Initialize(string apiKey, string appName,
+		protected override void OnBeforeNotify(Report report)
+		{
+			var ex = report.OriginalException;
+			var errorReport = new ErrorReport(ex);
+			if (!IsInitialized || errorReport.ShowModal(Parent))
+			{
+				base.OnBeforeNotify(report);
+				return;
+			}
+
+			report.Ignore();
+		}
+
+		public static void Initialize(string apiKey, string appName,
 			string[] args = null, string launcherVersion = null, object parent = null,
 			[CallerFilePath] string filename = null)
 		{
 			Client = new ExceptionLoggingUI(apiKey, appName, args, launcherVersion, parent, filename);
-			return Client;
+			CommonInitalize();
 		}
 
 	}